Worked on Ninja Tournament
Only other thing added to Ninja Tournament is rewards/cash prizes for places other than first.
Also, a fee should be added to enter a tournament. e.g. must pay 6,250 pokelets to enter tournament
6,250 because if there are 16 "players" and cash prize is 100,000 each player pays fee and winner takes their pokelets

Another thing, the format for all tournaments is the same, the only thing that is different would be the "bonuses" for the computer like so that each round is harder, and the fee and cash winnings would be higher. I was thinking that we could put in 2-3 methods for those and only have one tournament method that uses other methods to calculate the fee and the cash prizes and stuff